requirements.txt
asgiref==3.7.2
asttokens==2.4.1
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
icecream==2.1.3
idna==3.7
Naked==0.1.32
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
six==1.16.0
snakeviz==2.2.0
sqlparse==0.5.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
ultraimport==0.0.7
urllib3==2.2.1